requirements.txt
flask
sqlalchemy
flask-sqlalchemy>=3.0.0
openai>=1.0.0
PyYAML>=6.0
pandas>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
chardet>=5.0.0
pytest>=7.0.0
requests>=2.28.0
